uftrace: Remove "+" prefix in --time-range option
-r/--time-range is used to specify time range either timestamp or
elapsed time.  The given data is parsed as timestamp by default or
parsed as elapsed time when it has "+" prefix and <time_unit> postfix,
for example '+100ms'.

But it'd be better to remove "+" prefix to identify whether it is
timestamp or elapsed time.  It can automatically be identified only with
<time_unit> postfix.

So this patch removes "+" prefix for time range parsing and also updates
relevant man pages.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>