uftrace: Supplement help message for a time-range option
The user can check a time-range usage with help message as below

    $ uftrace --help | grep time-range
      -r, --time-range=TIME~TIME Show output within the TIME range only

But I think the user can be confusing
because they can't know what does TIME mean
e.g. duration, timestamp, elapsed time etc.

Sure, `uftrace replay` man pages contain this
but I think it is better that there is the mention
not only in man page but also help message.

So I modify the help message for time-range.

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>