record: Support a float value for time filter
We can use --time-filter or -t option
to filter functions by time as below.

    $ uftrace -t 1us ./a.out

But currently we can't use a float value for the threshold
because uftrace only handle the value as uint64_t type.
So fix codes to enable the user to use a float value for it e.g.

    $ uftrace -t 0.3us ./a.out

But the limited number of digits before and after decimal point
is three digits by default (In timestamp, it's 9).

The new parse_time() handle digits before decimal point
in accordance with a particular unit (e.g. us, ms, s).
(The digits means a number of digits before decimal point.)

For example,
If the digits is over considering a given unit, cut some digits by /=10
If the digits is short considering a given unit, append zeroes by *=10

Furthermore, the new parse_time() can handle the cases such as 0.001us
that contains zeroes before decimal point.

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>