record: Fix a bug in copy_to_buffer()
The commit 7605cd0 ("record: Fix a compiler warning") added a check for
return value of write().  But it can be called during flush_shmem_list()
which is called after stop_all_writers() so the thread_ctl[1] was
closed.  In this case the write fails with EBADF and it makes record fail.

Cc: Taeung Song <treeze.taeung@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>