mcount: Fix a symbol-resolving bug on ARM
When writing to rec buffer in record_ret_stack(), it should cast the
child address before shifting.  Otherwise it cuts off the MSB resulting
in hex address in replay with no symbols.

  $ uftrace hello
  # DURATION    TID     FUNCTION
    15.304 us [ 6364] | <3b4>();
              [ 6364] | <558>() {
    92.812 us [ 6364] |   <37c>();
   105.156 us [ 6364] | } /* <558> */

The addressses should be like 0x10XXX but it was missing.
Also update missing frstack->unused rename in the dead code.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>