replay: Escape a newline at the end of string
String arguments often end with a newline character.  In replay output,
those strings break alignment so it'd be better to escape them out.

For example, consider following case:

  $ uftrace -A printf@arg1/s hello
  hello world
  # DURATION    TID     FUNCTION
     2.177 us [ 3925] | __monstartup();
     0.720 us [ 3925] | __cxa_atexit();
              [ 3925] | main() {
     8.420 us [ 3925] |   printf("hello %s
   ");
     9.622 us [ 3925] | } /* main */

The newline character after '%s' was shown as is and broke the output.
Now it's displayed as escaped like in a source code.

  $ uftrace -A printf@arg1/s hello
  hello world
  # DURATION    TID     FUNCTION
     2.283 us [ 3937] | __monstartup();
     0.719 us [ 3937] | __cxa_atexit();
              [ 3937] | main() {
     8.812 us [ 3937] |   printf("hello %s\n");
     9.913 us [ 3937] | } /* main */

Signed-off-by: Namhyung Kim <namhyung@gmail.com>