build: Fix command-line variable assignment
When a value contains '=', the LHS appeared twice in the result.  For
example, it happens if user wants to set a default CFLAGS:

  $ ./configure CFLAGS=-D_FORTIFY_SOURCE=2

  $ grep CFLAGS .config
  override CFLAGS := -D_FORTIFY_SOURCE=-D_FORTIFY_SOURCE=2

This is because, it takes the shortest match string for option name when
parameter substitution happenes in the shell.  Instead, it needs to use
longest match using '%%'.  Now it shows the flag properly.

  $ grep CFLAGS .config
  override CFLAGS := -D_FORTIFY_SOURCE=2

Signed-off-by: Namhyung Kim <namhyung@gmail.com>