kernel: Save active event format files
The kernel trace can contain event records which might be enabled
regardless of uftrace.  In this case, it needs to know the event format
to parse the records correctly.

Fortunately, the events directory in tracefs provides the overall state
of enabled events.  So it can simply skip disabled events and can save
enabled events without iterating the whole hierarchy uncessarilly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>