mcount: Basic support for user events
There are some kind of trace events readily available on userspace like
systemtap SDT or LTTng UST.  Add support for those events to the uftrace
so that it can track such information.  It can be used with kernel trace
events seamlessly as it already added UFTRACE_EVENT record type.

Like kernel events, it has "provider:event" format and event id will be
assigned for each enabled event.  The event info will be saved into the
'events' file in the data directory.  The record command will save the
id in the record, and then replay will translate it to the event info.

The first target is the SDT which is available on most Ubuntu systems by
default.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>