arch/x86: Add basic support for SDT event
The SDT events are inserted with a single-type NOP (0x90), so it needs
to be replaced to an invalid opcode so that uftrace can catch the SIGILL
to save the event info.  And it modifies the value of the RIP register
to proceed with a next instruction after the opcode.

Currently it doesn't save arguments since it requires parsing the format
and some possible memory dereferences.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>