symbol: Introduce task_find_sym_addr()
The task_find_sym_addr() is similar to task_find_sym() but uses a raw
addr and timestamp.  This is to handle entries saved in the function
stack (fstack) for uftrace report or graph.

Also change is_kernel_address() to receive a pointer to symtabs since
the kernel base address not saved in the symtabs.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>