kernel: Filter out sys_clock_gettime() by default
It will be served by VDSO so normally it's not needed.  But some (old)
kernel (esp. ARM32) doesn't support VDSO so the output will be mixed
with sys_clock_gettime() when kernel tracing is used.  Those functions
are called from libmcount (uftrace) so should be hidden from users.

On x86, it should not change anything in the output.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>