configure: Tolerate dashes '-' in --long-options
For RPM or Deb packaging where a configure script is expected to be
compatible with autotools, long options may contain additional dashes
that shouldn't be evaluated as variable names. Instead dashes are
turned into underscores.

At the same time, protect the $val with 'simple quotes' as it is done
for positional arguments to avoid accidentally running commands when
values contain whitespaces. While this takes care of whitespaces, there
is still the possibility to run into a 'single quote' in the value but
this is less likely.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>