configure: Retire the '--long-option value' syntax
On Fedora or Red Hat systems, flags may be passed as long options
without a value. On Fedora, a %configure macro is available to
uniformly set up configure-based projects but it eats up the prefix:

./configure [...] --disable-dependency-tracking --prefix=/usr [...]

Since this notation is not even documented, the '--long-option=value'
notation can be enforced. Long options without a value are silently
ignored.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>