mcount: Fix a bug when longjmp() called twice
The longjmp can be called multiple times, but current code missed to
update the PLT GOT address for longjmp.  In fact it tried to update the
address, however the rstack was changed during longjmp(), its dynamic idx
was also changed.  And then it updated wrong address for longjmp.

Later another longjmp was called, it refers the wrong address and
resulted in a crash like below:

  $ uftrace longjmp2
  mcount: /home/namhyung/project/uftrace/libmcount/plthook.c:570:plthook_exit
   ERROR: <2> invalid dynsym idx: 65535
  # DURATION    TID     FUNCTION
     1.462 us [29196] | __monstartup();
     0.843 us [29196] | __cxa_atexit();
              [29196] | main() {
     3.626 us [29196] |   _setjmp();
              [29196] |   foo() {
              [29196] |     longjmp() {
     1.514 us [29196] |   } /* _setjmp */

  uftrace stopped tracing with remaining functions
  ================================================
  task: 29196
  [0] main

Signed-off-by: Namhyung Kim <namhyung@gmail.com>