mcount: Fix PLT hook segfault about 16-byte alignment on GCC7
If uftrace is built by GCC7, plthook_entry() has `movaps` insn.
So the stack pointer must be aligned on a 16-byte boundary.
This was handled by the commit 7b80ea8c2 ("mcount: Fix segfault on GCC7")

However, the commit didn't handle the below `-pg -mfentry` case:

  $ gcc -o hello -pg -mfentry hello.c

  $ uftrace ./hello
  child terminated by signal: 11: Segmentation fault
  # DURATION    TID     FUNCTION
     1.741 us [14660] | __monstartup();
     1.519 us [14660] | __cxa_atexit();

So fix it, this situation was mentioned on #100.

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>