mcount: Do not hook PLT of mcount()
The mcount is handled by libmcount directly and should not be shown to
users.  So skip it from PLT hooking in the first place.  It also helps
performance since it was executing meaningless PLT hook for every mcount
calls.

I hoped to apply same for __cyg_profile_func_{enter,exit} too but it
led to errors..

    CC FPIC  libmcount/plthook.op
  /home/namhyung/project/uftrace/libmcount/plthook.c: In function ‘skip_plt_functions’:
  /home/namhyung/project/uftrace/libmcount/plthook.c:78:3:
  error: built-in function ‘__cyg_profile_func_enter’ must be directly called
     SKIP_FUNC(__cyg_profile_func_enter),
     ^~~~~~~~~

Signed-off-by: Namhyung Kim <namhyung@gmail.com>