test: Add more test cases for setjmp/longjmp
The setjmp/longjmp is tricky to support non-local goto.  So it needs to
save and restore rstacks when setjmp/longjmp is called.  These cases
test if a program has more than one pair of them.

Note that the longjmp3 testcase gets segfault for -finstrument-functions
but it seems not related to uftrace though (it crashes without uftrace).

Signed-off-by: Namhyung Kim <namhyung@gmail.com>