mcount: Remove unnecessary mcount_arch_parent_location()
The mcount_arch_parent_location() is to calculate a correct parent
location for THUMB-compiled programs on a ARM system.  So it does
nothing on x86 can caused a tiny performance impact.  As mcount_entry()
is a very hot path, reduce the meaningless call can help a little.

Change it to static inline for x86 so that it can optimized out.  And
define FIX_PARENT_LOC macro for ARM so that it can be called if defined.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>