filter: Check libc++.so when displaying std::string
Since the current std::string display implementation is dependent on
libstdc++.so, it may crash if other std::string implenmation is used
such as libc++.so.

So it is required to check if libc++.so is used when displaying
arg/retval in std::string format.

In this case, this patch does not print arg/retval in std::string
format and just makes an warning as follows:

  WARN: std::string display for libc++.so is not supported.

Suggested-by: Namhyung Kim <namhyung@gmail.com>
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>