record: Support python script for function entry and exit
This patch supports python binding for uftrace entry and exit so that one
can easily add a python script to hook function entry and exit and do
whatever they want.

The following script implements "uftrace_entry" and "uftrace_exit".

  def uftrace_entry(args):
      print("entry_addr : " + hex(args["entry_addr"]))
      print("ret_addr   : " + hex(args["ret_addr"]))
      return entry_addr

  def uftrace_exit(args):
      if retval:
          print("return value : " + hex(args["retval"]))
      return args["ret_addr"]

The above python functions "uftrace_entry" and "uftrace_exit" are
executed in the entry and exit of C/C++ functions that are traced.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>