arch/x86: remove needless adjustment the offset for child func addr
The commit 854abed ("Fix filtering on x86_64") subtracts
the offset from the start of child function which is 9 bytes
(1 for push; 3 for mov; 5 for call). But we don't need to do it anymore,
because match_ip() checks an address between the start address
of function to be filtered and it's the end address.

Furthermore, the calculation adjusting the offset is not perfect.
For examples, if the child address is 0x400693 on hello(),

 Dump of assembler code for function hello:
                0x400686 <+0>:	push   %rbp
                0x400687 <+1>:	mov    %rsp,%rbp
                0x40068a <+4>:	sub    $0x10,%rsp
                0x40068e <+8>:	callq  0x400520 <mcount@plt>
  child addr => 0x400693 <+13>:

We can get 0x40068a(i.e 0x400693 - 0x9), not 0x400686 (i.e. correct hello()
function address) by this calculation.
So remove it.

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>