mcount: Clean up trace-on/off trigger handling
The mcount_enabled flag which was set by the trace-on/off trigger is to
control recording.  Unlike filters, it doesn't prevent execution of
mcount functions and just skips recording.  This is because it wants to
see proper function call stack even if enabled anytime.

Move the code to make it clearer (I hope).

Signed-off-by: Namhyung Kim <namhyung@gmail.com>