record: Add script processing step in record time
This patch adds script processing step in record time so that user can
do using supported scripts.  It implements a general way to run script
processing in each function entry and exit.  Python is the only target
script as of now.

The 'simple.py' script implements "uftrace_entry" and "uftrace_exit"
and those functions are executed in each entry and exit of C/C++
functions that are traced.

  $ uftrace -S scripts/simple.py -F main tests/t-abc
  entry : main()
  entry : a()
  entry : b()
  entry : c()
  entry : getpid()
  exit  : getpid()
  exit  : c()
  exit  : b()
  exit  : a()
  exit  : main()
  # DURATION    TID     FUNCTION
              [30960] | main() {
              [30960] |   a() {
              [30960] |     b() {
              [30960] |       c() {
     2.247 us [30960] |         getpid();
     8.948 us [30960] |       } /* c */
    13.011 us [30960] |     } /* b */
    18.580 us [30960] |   } /* a */
    41.977 us [30960] | } /* main */

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>