arch/aarch64: Implement mcount_arch_undo_bindnow()
When "bind-now" is set, the binary doesn't have PLT (other than PLT0)
for the external functions.  However, in order to trace those functions,
it needs to have PLT so that the plt_hooker can be called.

Fortunately, PLT0 in AArch64 has the code to setup required info for
plt_hooker so that we can simply update the addresses in GOT to point
the PLT0 without any trampoline.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>