mcount: Introduce mcount_arch_plthook_addr()
The mtd.plthook_addr is an (original) address in PLTGOT to be restored
after dynamic resolver sets the real address.  It needs to be set back
to the address in order to trace PLT calls.

But adding single variable cannot support nested calls, and it can be a
problem in some cases.  Normally PLT original address can be calculated
using PLT index so add the mcount_arch_plthook_address() to get it in an
arch-specific way.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>