mcount: Eliminate plthook_guard
The plthook_guard is to prevent nested PLT calls since it had to keep an
original PLT address to be restored.  However it doesn't need to keep
the address as it can be calculated using mcount_arch_plthook_addr().
As we already removed the plthook_addr member, let's remove
plthook_guard as well.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>