dump: Show schedule event in chrome and flame-graph
Treat the linux:schedule event as function so that we can see it in the
output.  But a newly created thread starts with a sched-in event and it
confuses the output since it's unbalanced.

Use task->timestamp_last to detect the first sched-in event.  This also
fixes missing time range check in replay dump mode.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>