mcount: Preserve return address after exception
The __cxa_end_catch() will be called after an exception was handled by
C++ runtime.  The uftrace recovers rstack based on the current frame
pointer value and write unwound stack frames.

But it might corrupt current function's return address.  And it looks
like that mcount_rstack_restore() at the beginning of the function can
overwrite it with previous rstack's value if it has same parent
location (it usually is).  And it also needs to update the return
address in the rstack too due to the mcount_rstack_reset() below.

This fixes the crash in the exception test case.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>