test: Update C++ exception test case
The empty foo() and bar() might be optimized out so add a volatile
expression to suppress it.  And filter out the personality routines as
they can be vary in the output.  Also change the output to remove some
functions but keep them in the fixup() routine to be backward compatible.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>