trigger: Support argspec in -T option
The -A and -R options are treated differently because it needs to
propagate the argspec to uftrace replay (and others) so that they can
know how to parse the data.  Specifying arguments (and return value) in
the -T option is not supported since it lacks such info.

This patch fixes the problem by saving those info to the file header.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>