trigger: Fill argument and return value spec automatically
When user requests argument and/or return value without explicity
specification, use the built-in auto-args info to fill proper argument
and/or return value spec automatically.

It's possible to not having auto-args for a given function.  For that
case ignore the trigger flag if spec list is empty.  Change add_filter()
to return whether it actually adds a entry or not.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>