info: Use dynamically allocated buffer in cmd-info.c
Claudiu Balogh reported that uftrace cannot read data that has more than
1000 threads.  It's because the current implementation stores the
task list in a fixed size buffer (4096).

So this patch makes the buffer to be allocated dynamically to make it
work for any size of task list.  The same change is applied to
read_arg_spec because argspec can also be too long to put into the fixed
size buffer.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>