plthook: Make x86's mcount_arch_plthook_addr() as default
It seems that "weak" attribute has some bug on some old toolchains (like
in ubuntu 10.04).  Move the body of x86 implementation of the function
into the weak function and remove the x86 code in order to work around
the problem.

Other weak funcitons (and other arch) will be still broken but as they
are for some advanced usages, I won't fix them all.  It's too old.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>