build: Add install-elfutils.sh to build elfutils manually
Sometimes it'd be better to provide full compilation method especially
in cross compile environment or in some systems that cannot install
libelf properly for some reasons.

This adds "misc/install-elfutils.sh" to download and compile elfutils.
The usage is as follows:

  $ export CROSS_COMPILE=arm-linux-gnueabi-
  $ export ARCH=arm
  $ export CFLAGS="-march=armv7-a"
  $ ./misc/install-elfutils.sh --prefix=`pwd`/build.arm
  $ ./configure --prefix=`pwd`/build.arm --with-elfutils=`pwd`/build.arm

  $ make
  $ make install

Some of environmental variables and prefix is better to be same both for
install-elfutils.sh and configure.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>