script: Add a way to flush python internal buffer at fork
Sometimes, when fork is called in the target program, python output is
printed twice because its buffer is not flushed when creating a new
process.

This patch adds a preparation routine so that some additional actions
can be taken before a new process is created using fork.  And it uses
the routine to flush python interal buffer to avoid such problem by
calling "sys.stdout.flush()".

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>