Modify macro identifiers of all header files
There are two cases to modify macro identifiers in header files.

1. Following naming convention of the C++ language standard
It's not recommended to begin with an underscore followed by an uppercase letter, because it can be reserved to the implementation for any use. So it is better to avoid using leading or trailing underscores in the name of the header guard.

2. Replacing FTRACE with UFTRACE
FTRACE was changed to UFTRACE by @namhyung's requirement. (https://github.com/namhyung/uftrace/issues/187)

* utils/compiler.h
* utils/filter.h
* utils/fstack.h
* utils/symbol.h
* utils/utils.h

Signed-off-by: Sangwon Hong <qpakzk@gmail.com>