doc: Add the default value of '-f' option
Refereneced from the FIELD section of `man uftrace-replay`, the
default value of '-f', '--output-fields=FIELD', is 'duration,tid',
so add this like that of the '--column-offest=DEPTH' said.
Users may infer that fields are seperated by comma(,) from this.

Parenthesize it with quote(') as saying in `man uftrace-replay`.

Now the `uftrace --help` shows followed:

    $ ufrace --help
    Usage: uftrace [OPTION...]
       	    [record|replay|live|report|info|dump|recv|graph|script]
            [<program>]
    uftrace -- function (graph) tracer for userspace
    ...
    -f, --output-fields=FIELD  Show FIELDs in the replay output (default:
                               'duration,tid')
    ...

Signed-off-by: Hansuk Hong <flavono123@gmail.com>