trigger: do not use memcpy() for copying trigger
if we use memcpy to copy the &iter->trigger,
depending on the size of the uftrace_trigger *tr
we will ride optimization path to use SSE inside memcpy.
it will change status of SIMD register like mmx.
therefore we lose original value of register.
so, we decide not to use memcpy.

Signed-off-by: HanbumPark <kese111@gmail.com>