symbol: Remove symtem libraries from skip_libs[]
The skip_libs array is to maintain a list of libraries don't need to be
saved.  It includes libmcount family and other base system libraries
like glibc.  But they are not actually skipped since it needs to compare
SONAME instead of filename.

Thinking it again, I think it's ok to keep since speed-up from skipping
them would be negligible and it could be used later.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>