perf: Add warrning message when not available
Honggyu reported that his system could not show schedule event.  Looking
at it, I found that the system set /proc/sys/kernel/perf_event_paranoid
to 3 which disables the perf event for non-root users entirely.

To reduce confusion, it'd be better a warning message for those cases.
With -v option, it provides what users to check.

Closes #239.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>