plthook: Handle corrupted rstack after vfork()
It seems there's a problem dealing with vfork() in gcc in that it
sometimes see a mtdp->idx of 0 which should not happen.  So far, it only
happenes when vfork() in parent returns immediately (for some reason).

More investigation is needed but just work around it by restoring vfork
rstack for now.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>