uftrace: Do not call exit() when fail to set color
Honggyu reported that some basic utilities on linux close standard file
streams (stdout, stderr) right before exiting.  But the destructor in
libmcount is called after that, so it ends up failing to show message in
pr_dbg() especially when setting color for the messages.

It was observed by running record command with -v option and/or uftrace
info command after that.  It should not affect the return value of the
child anyway.

Fixes #280.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>