arch/i386 : added macro to get current PC.
since we support PIC, sometimes our code needs the current program counter.
You can not directly reference a PC on an x86 32-bit architecture.
Therefore, we use the macro "__x86.get_pc_thunk.ax".
However, this macro does not exist in older versions of gcc.
So, we had hardcode that macro and add to common.S.

Signed-off-by: HanbumPark <kese111@gmail.com>