mcount: Use parent's frame address to recover from exception
After an exception was thrown, it needs to find rstack unwound by the
exception handler.  To do it work correctly, uftrace needs to compare
the frame pointer of current execution and rstack's parent location
(return address).

But __builtin_frame_address(0) just returns the frame address of current
function (child or even child's mcount) which is adjusted already.  For
mcount() we have parent_loc which points to return address of the parent
function so it can use it for the frame address.  But for cygprof() and
xray() it's not - so use child's frame address unfortunately.

To make the __builtin_frame_address(1) work (but it won't call directly
due to the warning it emits), build libmcount with frame pointer.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>