misc: Add madvise related prototypes for auto-args
The madvise() system call is used to give advice or directions to the
kernel so it can provide a good idea understanding how the memory range
will be used.  So it'd be better to add it to auto args list as well.

It includes posix_madvise() and posix_fadvice() for the similar purpose.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>