mcount: Ensure initialization of wrapper functions
It was reported that using Intel TBB (Thread Building Block) crashes
uftrace during initialization.  Looking into it reveals that the TBB
calls internal C++ exception handling functions to verify GCC's
implementation.  But for some reason, it was before the libmcount
initialization so real function pointer was not set yet.

So initialize the real pointers directly when it's called before the
mcount init code.

Fixes #307.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>