perf: Keep last perf index after reading
Do not reset the handle->last_perf_idx after reading since it can be
used to retrieve the perf reader.  The get_perf_record() will always
read the file contents after this change and it's ok that the only user
of it is do_dump_file() which requires the behavior.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>