uftrace: Use single match pattern
Currently uftrace matches filters and events differently.  It was due to
kernel's internal filtering only supports glob-like match.  But it can
be processed in user space before passing the filter/event string to the
kernel.  So it'd be better to use single match engine (regex or glob)
for the whole pattern matching.

This changes to use 'regex' match by default.  Later commit will make it
optional.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>