mcount: Make mcount_finish() thread-safe
If mcount_finish() is called while other thread is running, it could
cause crash since it releases internal resources.  Add a reference
counter (mcount_active) to keep the number of active threads and free
the resource if there's no other task running.

But it also adds ~20% of overhead due to atomic instructions (+ memory
barriers) in the fast path.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>