mcount: Fix race between mcount_exit() and mcount_finish()
Now the recursion guard ensures that mcount_exit() won't be called when
it sees FINISH trigger during mcount_entry() for the task.  But it's
possible to get the FINISH trigger in another thread when it's about to
start the mcount_exit().

In that case, mcount_guard_recursion() might see the finish flag and
free rstack in the mtd.  But it needs to see the original return address
to continue the normal execution.

To fix that, add force argument to ignore the finish flag and proceed to
the function exit so that it can free the rstack at the unguard routine
after getting the original address.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>