symbol: Use hard-coded PLT size
It seems that lld (LLVM linker) misses to set entry size of the .plt
section.  This resulted in repeated dynamic symbols since the current
code relied on the size to calculate the symbol address.

As we know its actual size, just use a hard-coded value to work around
the problem Honggyu reported in the recent v8 JS engine.

Fixes: #344

Signed-off-by: Namhyung Kim <namhyung@gmail.com>