arch/x86_64: Add 'inject.c' to inject shared object to process
To inject a shared object into the currently running process, need to
make the target process load the shared object. To do this, first,
use 'ptrace' to over-write the prebuilt assembly code to executable
segment in the target process. second, manipulate PC register to run
injected assembly-code.

The pre-written assembly code contains many traps in between. when
target process have been trapped, we can catch the trap signal and
write the desired value in the cpu register.

Using this mechanism, write the absolute path of the shared object
('libtrigger.so') to memory which be allocated, and then pass it as
argument value when calling '__libc_dlopen_mode' function. This allows
the desired shared object to be loaded by the target process.

Signed-off-by: HanbumPark <kese111@gmail.com>