mcount: Add 'trigger.c' which implementation of 'libtrigger.so'
'libtrigger.so' have been injected by uftrace using ptrace to load
'libmcount-dynamic.so'. when injecting 'libtrigger.so',
'__libc_dlopen_mode' was used but it could not load
'libmcount-dynamic.so' for unknown reason. for that reason, to load
the 'libmcount-dynamic.so' need 'libtrigger.so' which load shared
object by using 'dlopen'.

Signed-off-by: HanbumPark <kese111@gmail.com>