session: Don't create a map for main executable
A map file is created from libmcount which gives whole memory map of the
target process.  And then uftrace loads symbol tables using the maps to
construct the address space of the process when replaying.  But it also
loads symbol table of the main executable separately.  So it ends up
having duplicate map for main executable.

When loading maps from the map file, one might think that we could just
ignore the map file for main executable.  But it needs to read the map
to retrieve the base address.  Add symtabs->exec_base field to keep the
info and ignore the map when loading.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>