symbol: Do not load main symbol tables twice
The map file contains a map for main executable so load_module_symtabs()
tries to load it.  When recording is about to end, it writes all symbol
tables to the data directory but it will load the duplicate table one
from the executable directly and another from the map file.  The second
load will be discard since it already saved the symbol file.

Then there's no need to save symbol table from the executable directly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>