gcov: Add a script to generate code coverage report
This patch adds a convenient script to generate a html report for code
converage.  Since it uses 'lcov' so it has to be installed first.

The usage is as follows:

  $ ./misc/run-lcov.sh

It reads code coverage information and calls lcov and genhtml in order
with proper options.  Then it generates "coverage.html" directory that
contains code coverage report in html format.

It also accepts one argument to change the output directory name.

  $ ./misc/run-lcov.sh xxx

The generated directory name becomes "xxx" instead of "coverage.html".

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>