tui: Consolidate tui_window implementation
The TUI graph and report have implemented same operations for window
movements with a little bit difference.  Consolidate them as tui_window
so that it can make easier for later modifications.

As the first member of tui_graph is now tui_window, it cannot cast a
pointer to uftrace_graph into tui_graph anymore.  Therefore the struct
tui_graph_node is changed to contain a pointer to uftrace_graph instead.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>