gcov: Support code coverage for libmcount-*.so
This patch allows libmcount libraries to be compiled with code coverage
support option.

Since gcov related options are smart enough, the target program doesn't
have to be compiled with code coverage option but just with normal
tracable option such as -pg, -finstrument-functions or without such
flags.

The --coverage option is a synonym for -fprofile-arcs -ftest-coverage
(when compiling) and -lgcov (when linking) so it's used for simplicity.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>