session: Set base address of executable only once
With the commit 9b293742 ("mcount: Skip duplicate maps in
record_proc_maps()") it records a map info for each module but old data
still contains duplicate maps.  This makes the old data not able to
resolve symbols since it updated the exec_base twice (for the data
map).  Make sure only does it once to fix the problem.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>