tui: Support to find longest executed child in tui graph
In tui mode, 'c' key is very useful by collapsing nested children.
It'd also be very useful to provide a short key to find the child that
has the longest execution time among all the other children.

The 'l' key for the longest child performs as follows:

  1. unfold the current node
  2. find a child that has the longest execution time
  3. collapse all the other children except for the longest child
  4. move to the longest child

It'd be useful when finding the performance critical execution path.

Resolved: #398

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>