tui: Collapse children after finding parent with 'u' key
'u' key is used to find its parent node but it may be too far away from
the current node if there are many nodes in call graph.  It may be
difficult to identify the current node and its parent in a single screen.

So it'd be useful to collapse the children after finding the parent node.
This may come in handy when using it with `l' key that finds longest
executed child function.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>