tui: Move to the top of window when switching from session window
Because the previous commit changes the default focus of report window
based on the current position of grpah window, it may sometimes be
confusing to users to see report window when they switch to report from
session window.

HOME key can simply move the focus to the top but some of the laptops
including mine doesn't have HOME button. :(

So this patch makes the default position to the top in the report window
when it's switched from session window.

Likewise, the same thing consistently applies to graph and info windows.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>