mcount: Use 32-bit PLT index in the mcount rstack
For a large program, it might be possible to have more than 64K of PLT
calls.  In that case deciding PLT call by the index will be broken.  In
the other hand, we don't support filter depth more than 64K, just
exchange the type of filter depth field and dyn_idx.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>