utils: Verify given data path is store of uftrace data.
-. if the path exist that specified by user to store trace data, uftrace
will append '.old' suffix to its name.
-. when path move to name which have appended '.old' suffix, if that
path exist then uftrace will remove it.

In the above process, uftrace does not verify that the path specified is
a path created by uftrace to store uftrace data. it can be harmful
because user can be mistake while specify the path. so, there is need to
verify given path is a path created by uftrace.

from this patch, uftrace verify the uftrace signature from the info file
that exists in the given path.

Fixed: #353

Signed-off-by: Hanbum Park <kese111@gmail.com>