python: Build symbol table for python functions
It needs to map python function name to an address so that uftrace can
deal with it like C functions.  As lowest address won't be used by
native programs, just use simple counter (from 1) to save symbol
address.  The symbol table will be written as "python-symtab.sym" in the
data directory.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>