test: Update 136 dynamic test
On my Archlinux machine, it failed to build like below:

  cc1: error: -mnop-mcount is not implemented for -fPIC

The reason was the system enabled PIE build by default.  Adding -fno-pie
option fixed it but it caused a crash at runtime like below:

  ./t-abc: Symbol `getpid' causes overflow in R_X86_64_PC32 relocation
  ./t-abc: Symbol `atoi' causes overflow in R_X86_64_PC32 relocation
  Segmentation fault (core dumped)

It related to library call assuming PIE/PIC relocation for some reason.
So I ended up add '-fno-plt' too to workaround the bug but it requires a
change in uftrace command line since library calls are not traced
anymore.  So --no-libcall was added and the output is changed as well.

It maybe has to be restored when this is fixed.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>