doc: Introduce a slideshow document for uftrace
This introduces a slideshow document for uftrace.  It is written based
on remark.js, a simple, in-browser, Markdown-driven slideshow tool.

It's goal is to create a simple, animated document so that users can
easily understand what uftrace does without reading a long description
in text-based documents.

The document itself is very long because it contains a lot of
near-duplicated pages to implement animation that describes the
execution procedure intuitively.

The page can be opened with the link below:

  http://uftrace.github.io

Hopefully, this can be helpful to uftrace users and developers.

Link: https://remarkjs.com
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>