perf: Look-ahead a schedule event for time filter
It should not display schedule events when time filter is used and the
period is shorter than the filter value.  However handling it properly
seems not so easy since perf events are managed for each cpu while
schedule events are by tasks.

This patch just adds a simple and partial solution which assumes
subsequent schedule event will be in a same cpu.  It just peeks a next
event when it sees a sched-out event for matching sched-in event.  And
then compare with the time filter and remove the both events if needed.
Also it cannot apply time trigger for current (or parent) function.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>