perf: Handle COMM event properly
The uftrace info shows task comm name using update_perf_task_comm() but
it didn't work actually since task handles are not initialized yet.  The
check for task is to protect it from unrelated events and it can be done
by checking find_task() which is constructed in open_data_file().

It missed to set the 'more' bit in the rstack so sometimes it might miss
the comm name from the event.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>