mcount: Invalidate rstack from parent at fork
When child starts run it inherit parent's rstack which might have some
records not flushed yet (probably due to --no-libcall).  In that case
child will write the functions it didn't run in record_trace_data().

This caused a segfault because the timestamp of the records are before
task creation so it couldn't find a session.  We might be more smart
when finding sessions but it'd be nice to prevent such functions in the
first place.

Fixed: #503

Signed-off-by: Namhyung Kim <namhyung@gmail.com>