test: fix conflict of a concurrent build of same test binary
When running tests, same test binary will be compiled at the same time.
For instance, 004, 005, 006 tests are using `s-abc` as a test binary.
Running concurrent build will compete to build their own targets.
And during this condition, binary waiting for a test can be removed.

To solve, this commit creates temp directory with tempfile.mkdtemp()
And to this directory, test will build with abs path of source, and
test with absolute uftrace path.

Signed-off-by: Daniel T. Lee <danieltimlee@gmail.com>